Company basics done. Separate all other components into their own folders and inplemented the ApiService file for each
